Move c10::DeviceType to a standalone header file
Summary:
* Extract the DeviceType enum class into a standalone header file in the new torch/standalone directory
* Keep c10/core/DeviceType.h files for backward compatibility
* Keep c10/core/DeviceType.cpp for static variables used for custom backend name registration
* Keep DeviceType declared in the c10 namespace for backward compatibility, and create a new torch::standalone namespace to alias DeviceType.
